agent-builder-webhook-alloydb/requirements.txt
functions-framework==3.*
google-cloud-alloydb-connector[pg8000]==1.0.0
langchain-core==0.1.45
langchain-google-vertexai==1.0.1
SQLAlchemy==2.0.29
tabulate==0.9.0
